requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
celery
redis
APScheduler
jinja2
plotly
matplotlib
reportlab
pytest
